+++ src/types/BaseProps.tsx
@@ -1,4 +1,4 @@
-import { Effects } from "./Effects";
+import { Effects, Waveforms } from "./Effects";
 import { OverlayTypes } from "../hocs/OverlayTilingSprite";
 import { Animations } from "./Animations";
 
@@ -6,7 +6,7 @@ export interface PixiBaseSpriteProps {
   children?: React.ReactNode;
   uniqueId: string;
   src?: string;
-  
+
   startAt: number;
   endAt: number;
   visible: boolean;
@@ -52,6 +52,7 @@ export interface PixiBaseSpriteProps {
     };
     effect?: Effects;
     overlay?: OverlayTypes;
+    waveform?: Waveforms;
   };
   pointerdown?: () => void;
   pointerup?: () => void;