+++ src/types/Effects.tsx
@@ -10,3 +10,12 @@ export enum Effects {
   Vignette = "Vignette",
   None = "Normal",
 }
+
+export enum Waveforms {
+  NONE = "none",
+  EQUALIZER = "equalizer",
+  ROUNDBARS = "roundbars",
+  LED = "led",
+  RADIAL_SPECTRUM = "radial_spectrum",
+  LUMI_BARS = "lumi_bars",
+}